alle Crud metoder virker vist

+++ CommentService.Application/CommentCommand.cs
@@ -1,11 +1,11 @@
-﻿using Dapr.Client;
-using CommentService.Application.Interfaces;
+﻿using CommentService.Application.Interfaces;
 using CommentService.Domain;
 using CommentService.Domain.Interfaces;
+using Dapr.Client;
 using Microsoft.Extensions.Logging;
-using CommentService.Application.Interfaces;
 using Shared.Comment;
 using Shared.Comment.DTO_s;
+using static Google.Rpc.Context.AttributeContext.Types;
 
 namespace CommentService.Application
 {
@@ -15,34 +15,42 @@ namespace CommentService.Application
         private readonly ILogger<CommentCommand> _logger;
         private readonly ICommentRepository _commentRepository;
         private readonly IUnitOfWork _unitOfWork;
-        public CommentCommand(DaprClient daprClient, ILogger<CommentCommand> logger, ICommentRepository CommentRepository, IUnitOfWork unitOfWork)
+        private readonly ICommentStateModelCommand _commentStateModel;
+        public CommentCommand(DaprClient daprClient, ILogger<CommentCommand> logger, ICommentRepository CommentRepository, IUnitOfWork unitOfWork, ICommentStateModelCommand commentStateModel)
         {
             _daprClient = daprClient;
             _logger = logger;
             _commentRepository = CommentRepository;
             _unitOfWork = unitOfWork;
+            _commentStateModel = commentStateModel;
         }
 
-        async Task<CommentResultMessage> ICommentCommand.CreateComment(CreateCommentDTO dto)
+        async Task<CommentResultMessage> ICommentCommand.CreateComment(CreateCommentDTO dto, Guid authId)
         {
             var result = new CommentResultMessage
             {
-                Status = "Creating"
+                Status = $"Creating comment on post {dto.PostId}"
             };
             try
             {
                 await _unitOfWork.BeginTransactionAsync();
 
-                var comment = Comment.CreateComment(dto.Content, dto.PostId, dto.UserId);
+                var comment = Comment.CreateComment(dto.Content, dto.PostId, authId);
 
                 await _commentRepository.CreateComment(comment);
                 await _unitOfWork.CommitAsync();
 
-                await _daprClient.SaveStateAsync("statestore", comment.Id.ToString(), comment);
+                var commentState = _commentStateModel.Create(comment);
+                await _daprClient.SaveStateAsync("statestore", comment.Id.ToString(), commentState);
 
-                _logger.LogInformation("Comment for post {PostId} created successfully", dto.PostId);
                 result.CommentId = comment.Id.ToString();
                 result.Status = "Created";
+                result.PostId = dto.PostId.ToString();
+                result.AuthorId = authId.ToString();
+                result.StatusCode = 201;
+
+
+                _logger.LogInformation("Comment for post {PostId} created successfully", dto.PostId);
                 return result;
             }
             catch (Exception ex)
@@ -56,7 +64,7 @@ namespace CommentService.Application
         }
 
 
-        async Task<CommentResultMessage> ICommentCommand.UpdateComment(Guid commentId, UpdateCommentDTO dto)
+        async Task<CommentResultMessage> ICommentCommand.UpdateComment(Guid commentId, UpdateCommentDTO dto, Guid authId)
         {
             var result = new CommentResultMessage
             {
@@ -70,15 +78,27 @@ namespace CommentService.Application
                 var comment = await _commentRepository.GetCommentById(commentId);
                 if (comment == null)
                 {
+                    result.StatusCode = 404;
                     throw new Exception("Comment not found");
+                }
+                if (comment.AuthorId != authId)
+                {
+                    result.StatusCode = 403;
+                    throw new Exception("You are not authorized to update this user");
                 }
                 comment.UpdateComment(dto.Content);
                 await _commentRepository.UpdateComment(comment, comment.RowVersion);
                 await _unitOfWork.CommitAsync();
 
+                var commentState = _commentStateModel.Create(comment);
                 await _daprClient.SaveStateAsync("statestore", comment.Id.ToString(), comment);
 
+                result.CommentId = comment.Id.ToString();
                 result.Status = "Updated";
+                result.PostId = comment.PostId.ToString();
+                result.AuthorId = comment.AuthorId.ToString();
+                result.StatusCode = 200;
+
                 _logger.LogInformation("Comment {CommentId} updated successfully", commentId);
                 return result;
             }
@@ -92,7 +112,7 @@ namespace CommentService.Application
             }
         }
 
-        async Task<CommentResultMessage> ICommentCommand.DeleteComment(Guid commentId)
+        async Task<CommentResultMessage> ICommentCommand.DeleteComment(Guid commentId, Guid authId)
         {
             var result = new CommentResultMessage
             {
@@ -106,20 +126,27 @@ namespace CommentService.Application
                 var comment = await _commentRepository.GetCommentById(commentId);
                 if (comment == null)
                 {
+                    result.StatusCode = 404;
                     throw new Exception("Comment not found");
+                }
+                if (comment.AuthorId != authId)
+                {
+                    result.StatusCode = 403;
+                    throw new Exception("You are not authorized to update this user");
                 }
 
                 await _commentRepository.DeleteComment(comment, comment.RowVersion);
                 await _unitOfWork.CommitAsync();
 
-                var cachedComment = await _daprClient.GetStateAsync<Comment>("statestore", commentId.ToString());
-                if (cachedComment != null)
-                {
-                    await _daprClient.DeleteStateAsync("statestore", commentId.ToString());
-                }
+                await _daprClient.DeleteStateAsync("statestore", commentId.ToString());
+
+                result.CommentId = comment.Id.ToString();
+                result.Status = "Deleted";
+                result.PostId = comment.PostId.ToString();
+                result.AuthorId = comment.AuthorId.ToString();
+                result.StatusCode = 200;
 
                 _logger.LogInformation("Comment {CommentId} deleted successfully", commentId);
-                result.Status = "Deleted";
                 return result;
 
             }
